ingestion/requirements.txt
kafka-python==2.0.2
trino==0.334.0
python-dotenv==1.0.0
